requirements.txt
# fastapi
fastapi
uvicorn
pydantic[email]
# File
python-multipart
# JWT
python-jose[cryptography]
# Hash
passlib[bcrypt]
# SQL (Relational) Databases
sqlalchemy
# migragtion
alembic